% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_transitions.R
\name{counttrans.tree}
\alias{counttrans.tree}
\title{Count transitions as the result of simulation on a tree.}
\usage{
counttrans.tree(rowpatterns, rowtaxon, colpatterns, simseqs, leftwin = 0,
  shift = 0, cyclic = FALSE,
  bases = sort(unique(unlist(strsplit(rowpatterns, "")))))
}
\arguments{
\item{rowpatterns}{The patterns that index rows.}

\item{rowtaxon}{The taxon counted in rows.}

\item{colpatterns}{A data frame with column names equal to the tips counted in columns, and one row of patterns per row of the output.}

\item{simseqs}{A named list of (initseq, finalseq) pairs; 'finalseq' is used.}
}
\value{
A tuplecounts object.
}
\description{
Count transitions as the result of simulation on a tree.
}
\seealso{
counttrans, tuplecounts
}
