% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simseq.R
\name{rinitseq}
\alias{rinitseq}
\title{Sample a Random Sequence}
\usage{
rinitseq(seqlen, bases, basefreqs = rep(1/length(bases), length(bases)))
}
\description{
Return a random sequence, to be used for the initial state, for instance.  
bases can in fact be longer patterns.
Note that if bases are longer than 1 character,
 DOES NOT actually produce a sting with frequencies basefreqs.
}
