% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_simseq.R
\name{show.simseq}
\alias{show.simseq}
\title{Show Simulated Sequence}
\usage{
show.simseq(x, printit = FALSE, maxchar = min(nchar(x$initseq), 200),
  latex = FALSE)
}
\description{
Display the output of simulation by lowercasing or replacing with "." bases that didn't change
}
