% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simseq.R
\name{simseq.config}
\alias{simseq.config}
\title{Simualte From a Configuration}
\usage{
simseq.config(seqlen, tlen, config, ...)
}
\description{
Run simseq but from a config object, basically a named list
 note: probably want to run contextual::fill.default.config( ) on config first.
}

