% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simseq.R
\name{simseq}
\alias{simseq}
\title{Simulate Context-Dependent Evolution of a Sequence}
\usage{
simseq(seqlen, tlen, mutpats, mutrates, selpats = list(),
  selfactors = lapply(selpats, sapply, function(x) 1), selcoef = numeric(0),
  patlen, bases = c("A", "C", "G", "T"), initseq,
  basefreqs = rep(1/length(bases), length(bases)), count.trans = FALSE,
  only.num.events = FALSE, gmfile = NULL, simplify = TRUE,
  quiet = FALSE, ...)
}
\description{
Simulate a random sequence and evolve it with genmatrix, wrapping mutations around as needed.
 record transition counts if count.trans it TRUE (e.g. for debugging)
}
\details{
Computation of the generator matrix may be costly; so it can be stored for later use,
 and reloaded, by passing a file name in gmfile.  Note that this differs from other genmatrix'es stored,
 because we store a few more things, remove changes in the wings, and divide through rates to avoid overcounting.

IF only.num.events is TRUE, then JUST set things up, and estimate how many mutation events will be needed,
  and return this (to check before embarking on something ridiculous).
}
