% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_transitions.R
\name{counttrans}
\alias{counttrans}
\title{Count Transitions}
\usage{
counttrans(ipatterns, fpatterns, initseq = simseqs[["initseq"]],
  finalseq = simseqs[["finalseq"]], simseqs, leftwin = 0, shift = 0,
  cyclic = FALSE, bases = sort(unique(unlist(strsplit(ipatterns, "")))))
}
\value{
A tuplecounts object.
}
\description{
count number of times initseq matches ipatterns while finalseq matches fpatterns
  optionally, cyclical
if shift is nonzero, return a list  with the counts in each of (shift) frames
}
