% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simseq.R
\name{simseq}
\alias{simseq}
\title{Simulate Context-Dependent Evolution of a Sequence}
\usage{
simseq(seqlen, tlen, mutpats, mutrates, selpats = list(),
  selfactors = lapply(selpats, sapply, function(x) 1), selcoef = numeric(0),
  patlen, bases = c("A", "C", "G", "T"), initseq,
  basefreqs = rep(1/length(bases), length(bases)), count.trans = FALSE,
  only.num.events = FALSE, gmfile = NULL, simplify = TRUE,
  quiet = FALSE, ...)
}
\arguments{
\item{seqlen}{Length of the sequence in base pairs.}

\item{tlen}{Length of time to simulate for (may be a vector).}

\item{mutpats}{List of list of T-mers describing mutation patterns.}

\item{mutrates}{Vector of mutation rates.}

\item{selpats}{List of list of selection patterns.}

\item{selfactors}{List of vectors of relative selection coefficients.}

\item{selcoef}{Vector of selection coefficients.}

\item{patlen}{Maximum pattern length (determines size of genmatrix), safe to let it be automatically determined.}

\item{bases}{Character string of bases.}

\item{initseq}{List of initial sequences to start from, of the same length as tlen (if missing will be simulated).}

\item{basefreqs}{Frequencies to simulate initial sequences from (independently) if missing.}

\item{count.trans}{Whether to count up the number of transitions.}

\item{only.num.events}{Just count how many events to expect?}

\item{gmfile}{File where a precomputed generator matrix and associated other objects are found (NOTE: differs from genmatrix for inference!),
or will be saved to if the file doesn't exist.}

\item{simplify}{Strip out the list if only simulating one sequence?}

\item{quiet}{Be quiet?}

\item{...}{Additional arguments passed to contextual::makegenmatrix.}
}
\description{
Simulate a random sequence and evolve it with genmatrix, wrapping mutations around as needed.
 record transition counts if count.trans it TRUE (e.g. for debugging)
}
\details{
Computation of the generator matrix may be costly; so it can be stored for later use,
 and reloaded, by passing a file name in gmfile.  Note that this differs from other genmatrix'es stored,
 because we store a few more things, remove changes in the wings, and divide through rates to avoid overcounting.

IF only.num.events is TRUE, then JUST set things up, and estimate how many mutation events will be needed,
  and return this (to check before embarking on something ridiculous).
}

