% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simseq.R
\name{simseq.tree}
\alias{simseq.tree}
\title{Simulate on a Tree}
\usage{
simseq.tree(seqlen, config, ...)
}
\value{
A list of the simulated sequences in the same order as the tips,nodes of the tree
}
\description{
Simulate on a Tree
}
